Extract MemberDetails
